requirements.txt
Flask==2.3.3
numpy==1.24.2
tensorflow==2.13.0
Pillow==9.5.0
matplotlib==3.7.1
gunicorn==20.1.0
